pyproject.toml
[project]
name = "raytracegame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pygame-ce>=2.5.6",
]
